+++ projects/pip-webui2-behaviors/src/lib/selectable/index.ts
@@ -1,2 +1,3 @@
 export * from './selectable.component';
+export * from './selectable.directive';
 export { PipSelectableModule } from './selectable.module';