+++ projects/pip-webui2-behaviors/src/lib/selected/index.ts
@@ -1,2 +1,0 @@
-export { PipSelectedComponent } from './selected.component';
-export { PipSelectedModule } from './selected.module';