+++ projects/pip-webui2-behaviors/src/public_api.ts
@@ -7,4 +7,3 @@ export * from './lib/focused/index';
 export * from './lib/hotkeys/index';
 export * from './lib/infinite-scroll/index';
 export * from './lib/selectable/index';
-export * from './lib/selected/index';