+++ src/app/app.component.ts
@@ -1,13 +1,48 @@
-import { Component, ViewChild, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
-import { MatSidenav } from '@angular/material';
-import { MediaObserver, MediaChange } from '@angular/flex-layout';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { MediaObserver } from '@angular/flex-layout';
+import { MatSidenav } from '@angular/material/sidenav';
+import { TranslateService } from '@ngx-translate/core';
 import { PipHotkeysService } from 'pip-webui2-behaviors';
 import { PipThemesService, Theme } from 'pip-webui2-themes';
-import { TranslateService } from '@ngx-translate/core';
-
-import { AppTranslations } from './app.strings';
+import { combineLatest, Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 import { ExmapleListItem } from './examples-list/shared/ExampleListItem';
+import { AppTranslations } from './app.strings';
+
+type ObservableMap<T> = {
+  [P in keyof T]: Observable<T[P]>;
+};
+
+type ObservableOrAnyMap<T> = {
+  [P in keyof T]: Observable<T[P]> | T[P];
+};
+
+export function zip<L, R>(arr1: L[], arr2: R[]): [L, R][] {
+  return arr1.map((k, i) => [k, arr2[i]]);
+}
+
+export function unzip<L, R>(arr: [L, R][]): [L[], R[]] {
+  return arr.reduce(([l, r], [a, b]) => [[...l, a], [...r, b]], [[], []]);
+}
+
+function combineLatestMap<T>(sources: ObservableOrAnyMap<T>): Observable<T> {
+  const obs = {} as ObservableMap<T>;
+  const vals: Object = {};
+  Object.keys(sources).forEach(k => {
+    if (Observable.prototype.isPrototypeOf(sources[k])) {
+      obs[k] = sources[k];
+    } else {
+      vals[k] = sources[k];
+    }
+  });
+  const sourceEntries = Object.entries<Observable<any>>(obs);
+  const [sourceKeys, sourceValues] = unzip(sourceEntries);
+
+  return combineLatest(sourceValues).pipe(
+    map((values) => Object.assign({} as T, vals, Object.fromEntries(zip(sourceKeys, values))))
+  );
+}
 
 @Component({
   selector: 'app-root',
@@ -48,34 +83,19 @@ export class AppComponent implements OnInit {
     }
   ];
 
-  public themesLocalNames: any = {
-    'candy-theme': 'Candy',
-    'unicorn-dark-theme': 'Unicorn Dark',
-    'pip-blue-theme': 'Blue',
-    'pip-grey-theme': 'Grey',
-    'pip-pink-theme': 'Pink',
-    'pip-green-theme': 'Green',
-    'pip-navy-theme': 'Navy',
-    'pip-amber-theme': 'Amber',
-    'pip-orange-theme': 'Orange',
-    'pip-dark-theme': 'Dark',
-    'pip-black-theme': 'Black',
-    'bootbarn-warm-theme': 'Bootbarn Warm',
-    'bootbarn-cool-theme': 'Bootbarn Cool',
-    'bootbarn-mono-theme': 'Bootbarn Mono',
-    'mst-black-theme': 'MST Black',
-    'mst-black-dark-theme': 'MST Black Dark',
-    'mst-mono-theme': 'MST Mono',
-    'mst-orange-theme': 'MST Orange',
-    'mst-orange-dark-theme': 'MST Orange Dark',
-    'mst-elegant-theme': 'MST Elegant'
-  };
-
-  public listIndex = 0;
   public themes: Theme[];
+  public theme: Theme;
+  public languages = ['en', 'ru'];
+  public language: string;
+
+  public ctx$: Observable<{
+    activeMediaQuery: boolean,
+    mode: string
+  }>;
+  public listIndex = 0;
   public selectedTheme: Theme;
-  public activeMediaQuery: boolean;
-  public mode: string;
+  public activeMediaQuery$: Observable<boolean>;
+  public mode$: Observable<string>;
   public app = 'Behaviors';
   public url: string;
   public langs: string[] = [];
@@ -90,18 +110,26 @@ export class AppComponent implements OnInit {
     private pipHotkeys: PipHotkeysService
   ) {
 
-    this.themes = this.service.themes;
-    this.selectedTheme = this.service.selectedTheme;
+    this.themes = this.service.themesArray;
+    this.theme = this.service.currentTheme;
 
-    translate.setDefaultLang(this.selectedLang);
-    translate.use(this.selectedLang);
-    this.langs = translate.getLangs();
-    this.translate.setTranslation('en', AppTranslations.en, true);
-    this.translate.setTranslation('ru', AppTranslations.ru, true);
+    this.translate.addLangs(this.languages);
+    Object.entries(AppTranslations).forEach(e => this.translate.setTranslation(e[0], e[1], true));
+    this.translate.setDefaultLang('en');
 
-    media.media$.subscribe((change: MediaChange) => {
-      this.activeMediaQuery = change && change.mqAlias === 'xs' ? true : false;
-      this.mode = change && change.mqAlias === 'xs' ? null : 'side';
+    const browserLang = translate.getBrowserLang();
+    this.translate.use(browserLang.match(/en|ru/) ? browserLang : 'en');
+    this.language = this.translate.currentLang;
+
+    this.activeMediaQuery$ = media.asObservable().pipe((map(changes => changes && changes.some(c => c.mqAlias === 'xs'))));
+    this.mode$ = media.asObservable().pipe((map(changes => changes && changes.some(c => c.mqAlias === 'xs') ? null : 'side')));
+    this.ctx$ = combineLatest([
+      this.activeMediaQuery$,
+      this.mode$
+    ]).pipe(map(([activeMediaQuery, mode]) => ({ activeMediaQuery, mode })));
+    this.ctx$ = combineLatestMap({
+      activeMediaQuery: this.activeMediaQuery$,
+      mode: this.mode$
     });
 
     this.pipHotkeys.add({
@@ -113,6 +141,7 @@ export class AppComponent implements OnInit {
     });
     this.pipHotkeys.add({
       hotkey: 'ctrl+alt+f',
+      action: () => { this.onListItemIndexChanged(1); },
       navigationCommand: ['focused']
     });
   }
@@ -123,14 +152,14 @@ export class AppComponent implements OnInit {
     }));
   }
 
-  public changeTheme(theme) {
-    this.selectedTheme = theme;
-    this.service.selectedTheme = this.selectedTheme;
+  public changeTheme(theme: Theme) {
+    this.theme = theme;
+    this.service.selectTheme(theme.name);
   }
 
-  public changeLanguage(lang) {
-    this.selectedLang = lang;
-    this.translate.use(lang);
+  public changeLanguage(language: string) {
+    this.language = language;
+    this.translate.use(language);
   }
 
   public onListItemIndexChanged(index: number) {