+++ src/app/selectable-example/selectable-example.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { CommonModule } from '@angular/common';
-import { MatListModule } from '@angular/material';
+import { MatListModule } from '@angular/material/list';
 import { TranslateModule } from '@ngx-translate/core';
 import { PipSelectableModule } from 'pip-webui2-behaviors';